delete time component
