+++ frontend/components/DailyTabs.tsx
@@ -1,9 +1,15 @@
-import { SyntheticEvent, useState } from 'react';
+import { Suspense, SyntheticEvent, useState } from 'react';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import Box from '@mui/material/Box';
+import dynamic from 'next/dynamic';
 
-import TimeComponent from './TimeComponent';
+const EventListing = dynamic(() => import("./EventListing"), {
+  loading: () => <b>Loading...</b>,
+});
+
+
+// import TimeComponent from './TimeComponent';
 import { TabPanelProps, DailyTabs } from '@/interfaces/Components';
 
 function TabPanel(props: TabPanelProps) {
@@ -100,11 +106,38 @@ export default function DailyTabs({
 
                 if (events.length) {
                     return (
-                      <TimeComponent
-                        key={time}
-                        events={events}
-                        time={time}
-                      />
+                      <div key={time}>
+                        <h2 className="time-title">
+                          {time}
+                        </h2>
+                        <div className="event-list">
+                          <div className='event-listing listing-header'>
+                            <div className='flex-row'>
+                              <div>
+                                Title
+                              </div>
+                            </div>
+                            <div className='event-details'>
+                              <div className='tickets-column'>
+                                Tickets
+                              </div>
+                              <div className='duration-column'>
+                                Duration
+                              </div>
+                              <div className='cost-column'>
+                                Cost
+                              </div>
+                            </div>
+                          </div>
+                          {events.map((eventIndex: number) => {
+                            return (
+                              <Suspense key={eventIndex}>
+                                <EventListing key={eventIndex} eventIndex={eventIndex} />
+                              </Suspense>
+                            )
+                          })}
+                        </div>
+                      </div>
                     )
                 }
               }
